+++ crates/router/src/types/api/disputes.rs
@@ -1,0 +1,15 @@
+use masking::Deserialize;
+
+#[derive(Default, Debug, Deserialize)]
+pub struct DisputePayload {
+    pub amount: String,
+    pub currency: String,
+    pub dispute_stage: api_models::enums::DisputeStage,
+    pub connector_status: String,
+    pub connector_dispute_id: String,
+    pub connector_reason: Option<String>,
+    pub connector_reason_code: Option<String>,
+    pub challenge_required_by: Option<String>,
+    pub created_at: Option<String>,
+    pub updated_at: Option<String>,
+}